+++ component/include/pl_network_stream.h
@@ -13,7 +13,7 @@ namespace PL {
 class NetworkStream : public Stream {
 public:
   /// @brief Default read operation timeout in FreeRTOS ticks
-  static const TickType_t defaultReadTimeout = 300 / portTICK_PERIOD_MS;
+  static constexpr TickType_t defaultReadTimeout = 300 / portTICK_PERIOD_MS;
 
   /// @brief Creates a closed network stream
   NetworkStream() {}